Move container list to projectcontainers app
