+++ meinberlin/apps/dashboard2/urls.py
@@ -9,9 +9,6 @@ urlpatterns = [
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)/projects/$',
         views.ProjectListView.as_view(),
         name='project-list'),
-    url(r'^organisations/(?P<organisation_slug>[-\w_]+)/containers/$',
-        views.ContainerListView.as_view(),
-        name='container-list'),
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)/blueprints/$',
         views.BlueprintListView.as_view(),
         name='blueprint-list'),
