+++ meinberlin/apps/projectcontainers/views.py
@@ -1,3 +1,6 @@
+from django.views import generic
+
+from meinberlin.apps.dashboard2 import mixins as dashboard_mixins
 from meinberlin.apps.dashboard2.components.forms.views import \
     ProjectComponentFormView
 
@@ -20,3 +23,17 @@ class ContainerProjectUpdateView(ProjectComponentFormView):
 
     def validate_object_module(self):
         return True
+
+
+class ContainerListView(dashboard_mixins.DashboardBaseMixin,
+                        generic.ListView):
+    model = models.ProjectContainer
+    paginate_by = 12
+    template_name = 'meinberlin_projectcontainers/container_list.html'
+    permission_required = 'a4projects.add_project'
+    menu_item = 'project'
+
+    def get_queryset(self):
+        return super().get_queryset().filter(
+            organisation=self.organisation
+        )
